refactor(#377): extract share button to component
