+++ components/FixedHeader/FixedHeader.tsx
@@ -1,9 +1,10 @@
 import { Breadcrumbs } from "components/Breadcrumbs";
 import { Clock } from "components/Clock";
+import { ShareButton } from "components/ShareButton";
+import useShare from "hooks/useShare";
 import { CitiesList } from "lib/cities";
-import React, { useEffect } from "react";
-import { IoShareSocial } from "react-icons/io5";
-import styled, { useTheme } from "styled-components";
+import React from "react";
+import styled from "styled-components";
 
 const StyledBreadcrumbs = styled(Breadcrumbs)<{ hasShare: boolean }>`
   justify-content: ${({ hasShare }) => (hasShare ? "flex-start" : "center")};
@@ -41,33 +42,11 @@ const InfoColumn = styled.div`
   grid-template-columns: 1fr;
 `;
 
-const ShareButton = styled.button`
-  background-color: transparent;
-  border: none;
-  cursor: pointer;
-  line-height: 0;
-  margin: 0;
-  padding: 0;
-`;
-
 interface Props {
   cities: CitiesList;
 }
 function FixedHeader({ cities }: Props) {
-  const theme = useTheme();
-  const [hasShare, setHasShare] = React.useState(false);
-
-  useEffect(() => {
-    if ("share" in navigator) {
-      setHasShare(true);
-    }
-  }, []);
-
-  const share = () => {
-    navigator.share({
-      url: window.location.href,
-    });
-  };
+  const hasShare = useShare();
   return (
     <StyledFixedHeader>
       <Wrapper>
@@ -75,11 +54,7 @@ function FixedHeader({ cities }: Props) {
           <StyledBreadcrumbs cities={cities} hasShare={hasShare} />
           <StyledClock hasShare={hasShare} />
         </InfoColumn>
-        {hasShare && (
-          <ShareButton type="button" title="Compartir" onClick={share}>
-            <IoShareSocial color={theme.colors.secondaryDark} />
-          </ShareButton>
-        )}
+        <ShareButton />
       </Wrapper>
     </StyledFixedHeader>
   );
