+++ components/ShareButton/index.ts
@@ -1,0 +1,1 @@
+export { default as ShareButton } from "./ShareButton";
